+++ packages/react-native/src/evm/components/ConnectWallet.tsx
@@ -8,7 +8,7 @@ import {
   useSwitchChain,
   useWalletContext,
 } from "@thirdweb-dev/react-core";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import { ActivityIndicator, Animated, StyleSheet } from "react-native";
 import { ConnectWalletButton } from "./ConnectWalletFlow/ConnectWalletButton";
 import { ConnectWalletButtonProps } from "./ConnectWalletFlow/ConnectWalletButton";
@@ -16,6 +16,7 @@ import BaseButton from "./base/BaseButton";
 import Text from "./base/Text";
 import { useUIContext } from "../providers/ui-context-provider";
 import { ThemeProvider } from "../styles/ThemeProvider";
+import { SupportedTokens, defaultTokens } from "./SendFunds/defaultTokens";
 
 export type ConnectWalletProps = {
   /**
@@ -45,6 +46,25 @@ export type ConnectWalletProps = {
    * @default false
    */
   switchToActiveChain?: boolean;
+
+  /**
+   * Override the default supported tokens for each network
+   *
+   * These tokens will be displayed in "Send Funds" Modal
+   */
+  supportedTokens?: SupportedTokens;
+
+  /**
+   * Show balance of ERC20 token instead of the native token  in the "Connected" button when connected to certain network
+   *
+   * @example
+   * ```tsx
+   * <ConnectWallet balanceToken={{
+   *  1: "0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599" // show USDC balance when connected to Ethereum mainnet
+   * }} />
+   * ```
+   */
+  displayBalanceToken?: Record<number, string>;
 } & ConnectWalletButtonProps;
 
 export const ConnectWallet = ({
@@ -52,11 +72,14 @@ export const ConnectWallet = ({
   theme,
   buttonTitle,
   modalTitle,
+  modalTitleIconUrl,
   extraRows,
   hideTestnetFaucet,
+  displayBalanceToken,
   switchToActiveChain,
   termsOfServiceUrl,
   privacyPolicyUrl,
+  supportedTokens,
 }: ConnectWalletProps) => {
   const fadeAnim = useRef(new Animated.Value(0)).current;
   const address = useAddress();
@@ -80,6 +103,20 @@ export const ConnectWallet = ({
       setTheme(theme);
     }
   }, [setTheme, theme]);
+
+  const supportedTokensMemo = useMemo(() => {
+    if (!supportedTokens) {
+      return defaultTokens;
+    }
+
+    const tokens = { ...defaultTokens };
+    for (const k in supportedTokens) {
+      const key = Number(k);
+      tokens[key] = supportedTokens[key];
+    }
+
+    return tokens;
+  }, [supportedTokens]);
 
   return (
     <ThemeProvider theme={theme}>
@@ -115,12 +152,15 @@ export const ConnectWallet = ({
               detailsButton={detailsButton}
               extraRows={extraRows}
               hideTestnetFaucet={hideTestnetFaucet}
+              supportedTokens={supportedTokensMemo}
+              displayBalanceToken={displayBalanceToken}
             />
           )
         ) : (
           <ConnectWalletButton
             modalTitle={modalTitle}
             buttonTitle={buttonTitle}
+            modalTitleIconUrl={modalTitleIconUrl}
             theme={theme}
             termsOfServiceUrl={termsOfServiceUrl}
             privacyPolicyUrl={privacyPolicyUrl}