+++ packages/react-native/src/evm/components/ConnectWalletFlow/ChooseWallet/ChooseWalletContent.tsx
@@ -40,7 +40,8 @@ export const ChooseWalletContent = ({
     <View style={{ flexDirection: "column", maxHeight: MAX_HEIGHT }}>
       <ScrollView
         style={{
-          marginVertical: 16,
+          marginTop: 16,
+          paddingBottom: 16,
           paddingHorizontal: 16,
         }}
       >
@@ -74,6 +75,7 @@ export const ChooseWalletContent = ({
                   recommended={item.recommended}
                   onPress={() => onChooseWallet(item)}
                   mb={marginBottom}
+                  paddingVertical="xxs"
                 />
               )}
             </Box>