+++ packages/react-native/src/evm/components/MainModal.tsx
@@ -2,18 +2,19 @@ import { useModalState } from "../providers/ui-context-provider";
 import { ConnectWalletFlow } from "./ConnectWalletFlow/ConnectWalletFlow";
 import { Dimensions, StyleSheet, View } from "react-native";
 import { useMemo } from "react";
-import { CLOSE_MODAL_STATE } from "../utils/modalTypes";
 import { ThemeProvider } from "../styles/ThemeProvider";
 import { SessionRequestModal } from "./ConnectWalletDetails/SessionRequestModal";
 import { SessionProposalModal } from "./ConnectWalletDetails/SessionProposalModal";
 import { TWModal } from "./base/modal/TWModal";
 import Box from "./base/Box";
+import { useAppTheme } from "../styles/hooks";
 
 const MODAL_HEIGHT = Dimensions.get("window").height * 0.6;
 const DEVICE_WIDTH = Dimensions.get("window").width;
 
 export const MainModal = () => {
-  const { modalState, setModalState } = useModalState();
+  const theme = useAppTheme();
+  const { modalState } = useModalState();
 
   const { isOpen, isSheet } = modalState;
 
@@ -32,13 +33,9 @@ export const MainModal = () => {
     }
   }, [modalState.view]);
 
-  const onBackdropPress = () => {
-    setModalState(CLOSE_MODAL_STATE("MainModal"));
-  };
-
   return (
-    <ThemeProvider>
-      <TWModal isVisible={isOpen} onBackdropPress={onBackdropPress}>
+    <ThemeProvider theme={theme}>
+      <TWModal isVisible={isOpen}>
         {isSheet ? (
           <Box backgroundColor="background" style={styles.modal}>
             <View style={styles.contentContainer}>{view}</View>