+++ packages/react-native/src/evm/components/base/TextInput.tsx
@@ -1,15 +1,21 @@
 import { StyleSheet, TextInput as TextInputRN } from "react-native";
 import Box from "./Box";
 import { useTheme } from "@shopify/restyle";
+import { ReactNode } from "react";
 
 type TextInputProps = {
   containerProps?: (typeof Box)["arguments"];
   textInputProps?: TextInputRN["props"];
+} & {
+  rightElement?: ReactNode;
+  leftElement?: ReactNode;
 };
 
 export const TextInput = ({
   containerProps,
   textInputProps,
+  rightElement,
+  leftElement,
 }: TextInputProps) => {
   const theme = useTheme();
 
@@ -23,6 +29,7 @@ export const TextInput = ({
       pr="xs"
       {...containerProps}
     >
+      {leftElement ? leftElement : null}
       <TextInputRN
         style={{ ...styles.textInput, color: theme.colors.textPrimary }}
         returnKeyType={"done"}
@@ -31,6 +38,7 @@ export const TextInput = ({
         autoCorrect={false}
         {...textInputProps}
       />
+      {rightElement ? rightElement : null}
     </Box>
   );
 };
@@ -40,6 +48,5 @@ const styles = StyleSheet.create({
     textAlign: "left",
     flex: 1,
     height: 40,
-    paddingLeft: 5,
   },
 });