+++ packages/react/src/wallet/wallets/embeddedWallet/EmbeddedWalletOTPLoginUI.tsx
@@ -6,10 +6,10 @@ import { FadeIn } from "../../../components/FadeIn";
 import { OTPInput } from "../../../components/OTPInput";
 import { Spacer } from "../../../components/Spacer";
 import { Spinner } from "../../../components/Spinner";
-import { Container, ModalHeader } from "../../../components/basic";
-import { Button } from "../../../components/buttons";
+import { Container, Line, ModalHeader } from "../../../components/basic";
 import { Text } from "../../../components/text";
 import { Theme, fontSize } from "../../../design-system";
+import { Button } from "../../../components/buttons";
 
 type EmbeddedWalletOTPLoginUIProps = ConnectUIProps<EmbeddedWallet>;
 
@@ -22,6 +22,7 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
     useWalletContext();
 
   const [wallet, setWallet] = useState<EmbeddedWallet | null>(null);
+  const isWideModal = props.modalSize === "wide";
 
   const [verifyStatus, setVerifyStatus] = useState<
     "verifying" | "invalid" | "valid" | "idle"
@@ -30,8 +31,6 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
   const [sendEmailOtpStatus, setSendEmailOtpStatus] = useState<
     "sending" | "sent" | "error"
   >("sending");
-
-  const recoveryCodeRequired = false;
 
   const sendEmail = useCallback(async () => {
     setOtpInput("");
@@ -55,15 +54,15 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
     }
   }, [createWalletInstance, email, props.walletConfig]);
 
-  const handleSubmit = () => {
-    if (!sendEmailOtpStatus || otpInput.length !== 6) {
+  const handleSubmit = (otp: string) => {
+    if (sendEmailOtpStatus !== "sent" || otp.length !== 6) {
       return;
     }
 
-    verifyCodes();
+    verifyCodes(otp);
   };
 
-  const verifyCodes = async () => {
+  const verifyCodes = async (otp: string) => {
     setVerifyStatus("idle");
 
     if (!wallet) {
@@ -76,7 +75,7 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
       await wallet.connect({
         loginType: "headless_email_otp_verification",
         email,
-        otp: otpInput,
+        otp,
       });
 
       setConnectedWallet(wallet);
@@ -98,8 +97,10 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
   }, [sendEmail]);
 
   return (
-    <Container fullHeight flex="column" p="lg" animate="fadein">
-      <ModalHeader title="Sign in" onBack={props.goBack} />
+    <Container fullHeight flex="column" animate="fadein">
+      <Container p="lg">
+        <ModalHeader title="Sign in" onBack={props.goBack} />
+      </Container>
 
       <Container expand flex="column" center="y">
         <form
@@ -112,11 +113,11 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
               textAlign: "center",
             }}
           >
-            <Spacer y="xxl" />
+            {!isWideModal && <Spacer y="xl" />}
             <Text>Enter the OTP sent to</Text>
             <Spacer y="sm" />
             <Text color="primaryText">{email}</Text>
-            <Spacer y="xl" />
+            <Spacer y="xxl" />
           </div>
 
           <OTPInput
@@ -125,75 +126,82 @@ export const EmbeddedWalletOTPLoginUI: React.FC<
             value={otpInput}
             setValue={(value) => {
               setOtpInput(value);
-              setVerifyStatus("idle");
+              setVerifyStatus("idle"); // reset error
+              handleSubmit(value);
             }}
-            onEnter={handleSubmit}
+            onEnter={() => {
+              handleSubmit(otpInput);
+            }}
           />
 
           {verifyStatus === "invalid" && (
             <FadeIn>
-              <Spacer y="sm" />
-              <Container flex="row" center="x">
-                <Text size="sm" color="danger">
-                  Invalid OTP {recoveryCodeRequired ? "or recovery code" : ""}
-                </Text>
-              </Container>
+              <Spacer y="md" />
+              <Text size="sm" color="danger" center>
+                Invalid OTP
+              </Text>
             </FadeIn>
           )}
 
-          <Spacer y="xl" />
+          <Spacer y="xxl" />
 
-          {verifyStatus === "verifying" ? (
-            <>
-              <Spacer y="md" />
-              <Container flex="row" center="x">
-                <Spinner size="md" color="accentText" />
+          <Container px={isWideModal ? "xxl" : "lg"}>
+            {verifyStatus === "verifying" ? (
+              <>
+                <Container flex="row" center="x" animate="fadein">
+                  <Spinner size="lg" color="accentText" />
+                </Container>
+              </>
+            ) : (
+              <Container animate="fadein" key="btn-container">
+                <Button
+                  onClick={() => handleSubmit(otpInput)}
+                  variant="accent"
+                  type="submit"
+                  style={{
+                    width: "100%",
+                  }}
+                >
+                  Verify
+                </Button>
               </Container>
-              <Spacer y="md" />
-            </>
-          ) : (
-            <Button
-              onClick={handleSubmit}
-              variant="accent"
-              type="submit"
-              style={{
-                width: "100%",
-              }}
-            >
-              Verify
-            </Button>
-          )}
+            )}
+          </Container>
 
-          <Spacer y="lg" />
+          <Spacer y="xxl" />
 
-          {sendEmailOtpStatus === "error" && (
-            <>
-              <Text size="sm" center color="danger">
-                Failed to send OTP
-              </Text>
-              <Spacer y="md" />
-            </>
-          )}
+          {!isWideModal && <Line />}
 
-          {sendEmailOtpStatus === "sending" && (
-            <Container
-              flex="row"
-              center="both"
-              gap="xs"
-              style={{
-                textAlign: "center",
-              }}
-            >
-              <Text size="sm">Sending OTP</Text>
-              <Spinner size="xs" color="secondaryText" />
-            </Container>
-          )}
+          <Container p={isWideModal ? undefined : "lg"}>
+            {sendEmailOtpStatus === "error" && (
+              <>
+                <Text size="sm" center color="danger">
+                  Failed to send OTP
+                </Text>
+                <Spacer y="md" />
+              </>
+            )}
 
-          {sendEmailOtpStatus === "sent" && (
-            <LinkButton onClick={sendEmail} type="button">
-              Resend OTP
-            </LinkButton>
-          )}
+            {sendEmailOtpStatus === "sending" && (
+              <Container
+                flex="row"
+                center="both"
+                gap="xs"
+                style={{
+                  textAlign: "center",
+                }}
+              >
+                <Text size="sm">Sending OTP</Text>
+                <Spinner size="xs" color="secondaryText" />
+              </Container>
+            )}
+
+            {sendEmailOtpStatus === "sent" && (
+              <LinkButton onClick={sendEmail} type="button">
+                Resend OTP
+              </LinkButton>
+            )}
+          </Container>
         </form>
       </Container>
     </Container>
@@ -206,6 +214,7 @@ const LinkButton = styled.button<{ theme?: Theme }>`
   font-size: ${fontSize.sm};
   cursor: pointer;
   text-align: center;
+  font-weight: 500;
   width: 100%;
   &:hover {
     color: ${(p) => p.theme.colors.primaryText};